refactor: Travelogue domain refactoring
- 임시저장 기능 MySQL 의 필드로 관리하는 것이 아닌 Redis와 같은 외부 저장소 사용 예정
- 필요없는 필드와 메서드 수정중
- 임시저장 및 임시저장 조회 관련 코드는 생성 및 조회 코드와 중복이라 제거

+++ src/main/java/shop/zip/travel/domain/post/travelogue/dto/TravelogueSimple.java
@@ -6,7 +6,7 @@ public record TravelogueSimple(
 		Long travelogueId,
 		String title,
 		Period period,
-		Long totalCost,
+		int totalCost,
 		String country,
 		String thumbnail,
 		String memberNickname,
